\name{exanet}
\alias{exanet}
\docType{data}

\title{Example gene regulatory network for testing purposes only}

\description{
  The network consists of 814 genes and 1,103 interactions.
}

\usage{
  data(exanet)
}

\format{
  An undirected weighted igraph object with 814 vertices and 1,103 egdges.
}

%\details{

%}

%\source{

%}

\references{
   de Matos Simoes R, Emmert-Streib F., "Bagging statistical network inference from large-scale gene expression data" PLoS One. 2012;7(3):e33624. Epub 2012 Mar 30. PubMed PMID: 22479422; PubMed Central PMCID: PMC3316596.
}
 
\examples{
  data(expmat)
}


\keyword{datasets}
