\name{expmat}
\alias{expmat}
\docType{data}

\title{Test gene expression dataset}

\description{
  The dataset is test dataset of a gene expression matrix
  with 100 genes and 100 samples
}
\usage{
data(expmat)
}

\format{
  A matrix with 100 observations and 100 variables.
}

%\details{

%}

%\source{

%}
\references{
   de Matos Simoes R, Emmert-Streib F., "Bagging statistical network inference from large-scale gene expression data" PLoS One. 2012;7(3):e33624. Epub 2012 Mar 30. PubMed PMID: 22479422; PubMed Central PMCID: PMC3316596.
 }
 
\examples{
  data(expmat)
}


\keyword{datasets}
